requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2
django-bootstrap3==14.0.0
django-bootstrap4==2.1.1
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-mail-templated==2.6.5
django-registration==2.4.1
django-six==1.0.4
django-summernote==0.8.11.6
django-welcome==0.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.6.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.3
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
